requirements.txt
fastapi==0.95.0
SQLAlchemy==2.0.9
alembic==1.12.1
pydantic==1.10.7
aiologger==0.7.0
python-dotenv==1.0.0
asyncpg==0.27.0
psycopg2-binary==2.9.6
uvicorn==0.21.1
aiohttp==3.8.6
beautifulsoup4==4.12.0
lxml